requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sentence-transformers
torch
torchvision
# torchaudio # Может понадобиться для некоторых конфигураций torch/CLIP
openai-clip
Pillow
python-multipart # Для загрузки файлов
requests # Для загрузки изображений по URL (если нужно)